Made the staging bucket name location-specific
The buckets that we create are regional.
This prevents errors when some service required regional bucket.
E.g. "FailedPrecondition: 400 The Cloud Storage bucket of `gs://...` is in location `us`. It must be in the same regional location as the service location `us-central1`."
We are making the bucket name region-specific since the bucket is regional.