docs: resurface googleapis.dev and prediction docs (#1724)
I've mistakenly broken some of the docs for types links on googleapis.dev. Fixing the docs structure so the docs are showing again. If you run `nox -s docs` and take a look around the docs, they'll be back up.

I've verified that prediction pages are back up on c.g.c.

- [x] Make sure to open an issue as a [bug/issue](https://togithub.com/googleapis/python-aiplatform/issues/new/choose) before writing your code!  That way we can discuss the change, evaluate designs, and agree on the general idea
- [x] Ensure the tests and linter pass
- [x] Code coverage does not decrease (if any source code was changed)
- [x] Appropriate docs were updated (if necessary)

Fixes #1722 🦕