fix: LLM - De-hardcoded the `max_output_tokens` default value for the `CodeGenerationModel`
The previously default value (128) was inconsistent with the service-side default values of the models: `code-bison` has 1024 and `code-gecko` has 64.
More so, the default value was out of range of the `code-gecko` model.

This CL fixes these issues.

The SDK now relies on the service-side default values when the user does not pass a parameter value explicitly.

What can change: When using the `code-bison` model, the default value of `max_output_tokens` effectively increases from 128 to 1024 (the current service-side default value).

PiperOrigin-RevId: 559266968