Bugfix: 0 is now considered a valid step number in log_time_series_metrics
Previously, when log_time_series_metrics is called without a step value it defaults to None. The function internally checked `if not step` and
in this case, determined a step number. The problem is that 0 is falsy and therefore the function turns step=0 into step=1 which is a problem if
the function is later called with step=1 (in that case the first logged values are overwritten)

Checking for `if step is None` fixes the problem.