chore: GenAI - Fixed and improved the function calling handling in the GenAI unit tests
The `_RESPONSE_FUNCTION_CALL_PART_STRUCT` was incorrect (the `fields`  field should not have been there). After fixing that, the function call returned by the mock GenAI service has to be fixed due to the proto-plus library bug.
Additionally, improved the `test_chat_function_calling` test to properly respond to the function call in a realistic manner (by calling a function with the arguments returned by the model).

PiperOrigin-RevId: 616977403