chore: Refactored and split the `VertexAiResourceNounWithFutureManager` class
The `VertexAiResourceNounWithFutureManager` class was inheriting from `VertexAiResourceNoun` and `FutureManager` classes.
However, some of the functionality that the `VertexAiResourceNounWithFutureManager` class was adding was unrelated to `FutureManager`. This functionality was moved to the intermediate `_VertexAiResourceNounPlus` class so that some other classes can inherit from it without inheriting the `FutureManager` public API surface.

The new `_VertexAiResourceNounPlus` class does not add any public methods.

PiperOrigin-RevId: 621952481