feat: Support custom predictor Docker image builds on non-x86 architectures (#2115)
* Enforce Linux `x86_64` Docker image builds

Under the assumption that only machine types with x86 processors are
supported for prediction and custom training.

* Extract Docker build platform arg to method parameter

While currently only x86 processors are supported for prediction and
custom training, this will allow users to control this behavior should
that ever change in the future.

Additionally, it allows users to, e.g., override the `TARGETOS`
component of the `TARGETPLATFORM`.

* Change platform default arg to `None`

To enforce the flag is set by users as opposed to providing a universal
default.
- See: https://github.com/googleapis/python-aiplatform/pull/2115#discussion_r1203125274

* Make platform configurable from `LocalModel.build_cpr_model()`

To enable the flag to be set by users (e.g., to build images on non-x86
architectures).

* Fix docstring for `platform` param

Resolves:
- https://github.com/googleapis/python-aiplatform/pull/2115#discussion_r1978490069
- https://github.com/googleapis/python-aiplatform/pull/2115#discussion_r1978490463

* Test platform parameter in `test_build_cpr_model_upload_and_deploy()`

Resolves:
- https://github.com/googleapis/python-aiplatform/pull/2115#pullrequestreview-2655739124

* Fix tests

Resolves (partially):
- https://github.com/googleapis/python-aiplatform/pull/2115#issuecomment-2698426950

* Test specifying platform in local model builds

Resolves (partially):
- https://github.com/googleapis/python-aiplatform/pull/2115#issuecomment-2698426950

* Test other platform strings in local model builds

---------

Co-authored-by: Chun-Hsiang Wang <chunhsiang@google.com>